SelectSkillをUseSkillに統合、Monobehaviourを減らす

+++ Assets/Script/Card/CardDefine/SkillComponent/Particular/ScriptableCoinSkill.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public abstract class ScriptableCoinSkill : ScriptableObject, ICoinSkill
 {
-    protected abstract void Skill(CardFacade dealer, Coin c, short n);
+    protected abstract void Skill(CardFacade dealer, Coin c, int n);
 
-    public SkillProcess CoinSkill(Coin coin, short n)
+    public SkillProcess CoinSkill(Coin coin, int n)
     {
         return new SkillProcess(
         (CardFacade dealer) => { Skill(dealer, coin, n); }
